Tank now rotates and moves in that direction. WORKS.

+++ src/GameNetwork.java
@@ -177,8 +177,8 @@ public class GameNetwork
             Point b = a.getLocation();
             mouseX = (int) b.getX();
             mouseY = (int) b.getY();
-            tankX = game.getLocalTank().xPos();
-            tankY = game.getLocalTank().yPos();
+            tankX = (int) game.getLocalTank().xPos();
+            tankY = (int) game.getLocalTank().yPos();
             String inp = "mX"+mouseX+"mY"+mouseY+"tX"+tankX+"tY"+tankY;
             // convert the String input into the byte array.
             buf = inp.getBytes();
